いい感じに実装ができた！いくつか確認項目があるが、メールアドレス認証からユーザー情報登録が良さそうな実装になった.
