+++ docsthisbranch/Account-Feature-Evaluation.md
@@ -1,0 +1,99 @@
+# Account機能の評価レポート
+
+## StateManagement.mdの観点からの評価
+
+### ✅ 優れている点
+
+1. **状態の分類が明確**
+   - **Ephemeral UI state**: Zustand（フォーム、アップロード状態）
+   - **Remote server state**: TanStack Query（プロフィールデータ）
+   - **Persistent client state**: MMKV（hasCompletedProfileのみ）
+
+2. **Zustand規約の遵守**
+   - ファイル命名: `account-store.ts` ✅
+   - 型定義: `interface`を適切に使用 ✅
+   - middleware順序: `devtools → persist → immer` ✅
+   - 永続化制限: 設定のみを永続化（100KB以下） ✅
+
+3. **TanStack Query規約の遵守**
+   - queryKey: シリアライズ可能な配列 ✅
+   - staleTime/gcTime: 適切な設定値（5分/15分） ✅
+   - 楽観的更新の実装 ✅
+
+4. **Hook設計**
+   - 命名規則に従っている（`useProfileQuery`, `useCreateProfileMutation`） ✅
+   - セレクターによる最適化 ✅
+   - 統合フック（`useAccountProfile`）の提供 ✅
+
+### ⚠️ 改善可能な点
+
+1. **shallow比較の不要なインポート**
+   ```typescript
+   import { shallow } from 'zustand/shallow';
+   ```
+   - 最新のZustandでは不要（使用もされていない）
+
+2. **アクションセレクターの実装**
+   ```typescript
+   export const useAccountFormActions = () => 
+     useAccountFormStore((state) => ({
+       updateForm: state.updateForm,
+       // ... 全アクションを返す
+     }));
+   ```
+   - 無限ループの原因となったパターン（現在は個別取得で解決済み）
+
+## CentralizedStateManagement.mdの観点からの評価
+
+### ✅ 優れている点
+
+1. **一元管理の判断基準を満たしていない**
+   - Accountデータは主にプロフィール作成画面でのみ使用
+   - 他の機能からの依存が少ない
+   - **結論**: 個別管理が適切（現在の実装は正しい）
+
+2. **シンプルなProvider実装**
+   - StateManagerパターンを使用していない ✅
+   - 必要最小限のContext値 ✅
+   - 過度な抽象化を避けている ✅
+
+3. **クリーンアーキテクチャの遵守**
+   ```
+   account/
+   ├── application/      # Zustandストア
+   ├── domain/          # エンティティ
+   ├── infrastructure/  # APIサービス
+   └── presentation/    # UI層
+   ```
+
+4. **Blob管理の最適化**
+   - Blobデータをstoreに保持しない ✅
+   - useRefでの一時管理 ✅
+   - URIのみをstoreに保存 ✅
+
+### ⚠️ 注意点
+
+1. **AccountProviderの役割**
+   - 現在は主にプロフィール存在確認のみ
+   - 将来的にAccount関連機能が増えた場合は再評価が必要
+
+## 総合評価: A（優秀）
+
+現在のAccount機能の実装は、StateManagement.mdとCentralizedStateManagement.mdの原則に非常によく従っています：
+
+1. **状態管理の原則を正確に適用**
+2. **過度な一元管理を避け、適切な個別管理を選択**
+3. **パフォーマンスとメモリ使用量の最適化**
+4. **シンプルで理解しやすいコード構造**
+
+## 推奨事項
+
+1. **不要なインポートの削除**
+   - `shallow`のインポートを削除
+
+2. **将来の拡張性**
+   - Account機能が拡張される場合は、一元管理の必要性を再評価
+   - 現時点では個別管理が最適
+
+3. **ドキュメントの更新**
+   - CentralizedStateManagement.mdにAccount機能の実装例を追加
